move things to be cleaner
