+++ acq4/devices/PressureControl/__init__.py
@@ -1,2 +1,2 @@
 from .widgets import PressureControlWidget
-from .pressurecontrol import PressureControl
+from .device import PressureControl
