--- acq4/devices/SensapexPressureControl/__init__.py
@@ -1,1 +1,0 @@
-from .pressurecontrol import SensapexPressureControl
